As discussed with Abhineet, the histogram data needed for use by this template is not automatically created. Therefore, I had to modify the solution to Assignment 2 to create this histogram data for a test set of bear/bridge images.
Currently, I use 3 main scripts to do what needs to be done. The main.m script calls constructIMDB2.m (original, I know) 3 times to get 3 sets of data. One is for all images, one is for a set of positive images, and one is for a set of negative images, all of which exist in data/images. Once this is done, compareImages.m (exercise2.m with different hard coded data) may be called to do the object categorization which produces a set of images that it thinks contains the features that we are looking for.

Please note that I have hard coded in constructIMDB2.m to resize the images to 25% of their original size to aid in speed (at the cost of effectiveness). I have also hard coded the scripts for the 125 total images which I have testing with. Of course the runtime will take much longer when we use all images at 100% size (a matter of hours), but will hopefully give us a good showing of bear images.

What needs to be done? We need to make sure the program is doing what it is supposed to be doing, while I get an output of bear images, it could just be lucky. We need to alter the compareImages.m output to tell us how many bear images are found (and maybe the file names) rather than showing images, scores, and a graph. We need to run the program at 100% accuracy once these two things are done to get some results!