Last of several commits. SourceTree gave me an error that I couldn't push commits larger than 100MB so I did it in smaller chunks.
This last push completes the initial working prototype. Currently the program uses 235 input images (10 are positive). Main creates the data (attached), and compareImages.m determines which images are thought to be bear images. We need to improve on the bear detection by better specifying the features we would like to extract. This may be done by altering the files in myImages folder.